Add API endpoints for writing aggregates
Add API endpoint to append aggregates, and update the
aggregates-consumer to use the endpoint instead of writing directly to
the aggregates service's database.
Add API endpoint to upsert aggregates, and update the
reconciliation-worker to use the endpoint instead of writing directly to
the aggregates service's database.
Organize `main.go` to be less cluttered
Set up database migrations, using containerized `dbmate`, for
aggregates-db. This allows for applying migrations to a test database
for full integration testing of endpoints.